\name{FBFoptics-package}
\alias{FBFoptics-package}
\alias{FBFoptics}
\docType{package}
\title{
4x4 Optics Calculations
}
\description{
Package to calculate reflectivity and other parameters from Surface Plasmon Resonance (SPR) and Evanescent Wave Ellipsometry experiments in a prism configuration
}
\details{
\tabular{ll}{
Package: \tab FBFoptics\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-07-15\cr
License: \tab GPLv3\cr
}
Prepare a base object that describes the experimental configuration

}
\author{
Jonathan Bramble, School of Physics and Astronomy, University of Leeds

Maintainer: Jonathan Bramble <j.p.bramble@leeds.ac.uk>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
