% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Layer-class}
\alias{Layer}
\alias{Layer-class}
\title{An S4 class to represent a thin dielectric or metallic layer.}
\description{
An S4 class to represent a thin dielectric or metallic layer.
}
\section{Slots}{

\describe{
\item{\code{type}}{Denotes isotropic or anistropic layer, Currently unused argument, character}

\item{\code{d}}{Thickness of the layer in m}

\item{\code{eps}}{The relative permittivity of the material as a complex number}
}}

