% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{SPR-class}
\alias{SPR}
\alias{SPR-class}
\title{An S4 class to represent a SPR experiment.}
\description{
An S4 class to represent a SPR experiment.
}
\section{Slots}{

\describe{
\item{\code{points}}{The number of data points}

\item{\code{lambda}}{Wavelength of light}

\item{\code{n_entry}}{Refractive index of the entry medium, eg a Prism}

\item{\code{n_exit}}{Refractive index of the exit medium}

\item{\code{start_angle}}{Exterior starting angle in degrees}

\item{\code{end_angle}}{Exterior ending angle in degrees}
}}

